Use int task 1 0x15-file_io

+++ 0x15-file_io/1-create_file.c
@@ -13,7 +13,7 @@ int create_file(const char *filename, char *text_content)
 {
 	int fd;
 	int len;
-	ssize_t _write;
+	int _write;
 
 	if (filename == NULL)
 	{
